+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added: Performance improvements
 * Added: More debug mode messages
 
 ## [1.4.2] - 2017-08-29